Move metadata into its own file.
This makes no significant change, but attempts to make the code easier
to reason about. In the next steps the RefPointer code shouldn't know
about metadata explicitly at all.
